+++ inst/include/dust/dust.hpp
@@ -36,6 +36,7 @@ public:
     _n_threads(n_threads),
     _device_id(device_id),
     _rng(_n_particles_total, seed),
+    _errors(n_particles),
     _stale_host(false),
     _stale_device(true) {
 #ifdef __NVCC__
@@ -57,6 +58,7 @@ public:
     _n_threads(n_threads),
     _device_id(device_id),
     _rng(_n_particles_total, seed),
+    _errors(n_particles),
     _stale_host(false),
     _stale_device(true) {
 #ifdef __NVCC__
@@ -150,8 +152,13 @@ public:
     #pragma omp parallel for schedule(static) num_threads(_n_threads)
 #endif
     for (size_t i = 0; i < _particles.size(); ++i) {
-      _particles[i].run(step_end, _rng.state(i));
+      try {
+        _particles[i].run(step_end, _rng.state(i));
+      } catch (std::exception const& e) {
+        _errors.capture(e, i);
+      }
     }
+    _errors.report();
     _stale_device = true;
   }
 
@@ -240,12 +247,17 @@ public:
     #pragma omp parallel for schedule(static) num_threads(_n_threads)
 #endif
     for (size_t i = 0; i < _particles.size(); ++i) {
-      for (size_t t = 0; t < n_time; ++t) {
-        _particles[i].run(step_end[t], _rng.state(i));
-        size_t offset = t * n_state() * n_particles() + i * n_state();
-        _particles[i].state(_index, ret.begin() + offset);
+      try {
+        for (size_t t = 0; t < n_time; ++t) {
+          _particles[i].run(step_end[t], _rng.state(i));
+          size_t offset = t * n_state() * n_particles() + i * n_state();
+          _particles[i].state(_index, ret.begin() + offset);
+        }
+      } catch (std::exception const& e) {
+        _errors.capture(e, i);
       }
     }
+    _errors.report();
     return ret;
   }
 
@@ -450,6 +462,16 @@ public:
     return _shape;
   }
 
+  void check_errors() {
+    if (_errors.unresolved()) {
+      throw std::runtime_error("Errors pending; reset required");
+    }
+  }
+
+  void reset_errors() {
+    _errors.reset();
+  }
+
   std::vector<uint64_t> rng_state() {
     refresh_host();
     return _rng.export_state();
@@ -571,6 +593,7 @@ private:
   int _device_id;
   dust::pRNG<real_t> _rng;
   std::map<size_t, std::vector<data_t>> _data;
+  dust::openmp_errors _errors;
 
   std::vector<size_t> _index;
   std::vector<Particle<T>> _particles;
@@ -647,6 +670,7 @@ private:
       _shared = {pars.shared};
       initialise_device_data();
     }
+    reset_errors();
     update_device_shared();
     _stale_host = false;
     _stale_device = true;
@@ -688,6 +712,7 @@ private:
       }
       initialise_device_data();
     }
+    reset_errors();
     update_device_shared();
     _stale_host = false;
     _stale_device = true;