+++ inst/include/dust/utils.hpp
@@ -1,6 +1,8 @@
 #ifndef DUST_UTILS_HPP
 #define DUST_UTILS_HPP
 
+#include <cassert>
+
 namespace dust {
 namespace utils {
 
@@ -61,6 +63,15 @@ HOSTDEVICE T epsilon() {
   return epsilon_nvcc<T>();
 #else
   return std::numeric_limits<T>::epsilon();
+#endif
+}
+
+inline HOSTDEVICE void throw_message(const char * msg) {
+#ifdef __CUDA_ARCH__
+    printf("%s\n", msg);
+    assert(0);
+#else
+    throw std::runtime_error(msg);
 #endif
 }